Slim down property patch database
